pyproject.toml
[tool.poetry]
name = "running-api"
version = "0.1.0"
description = ""
authors = ["viniciussslima <viniciussslima999@gmail.com>"]
readme = "README.md"
packages = [{include = "running_api"}]

[tool.poetry.scripts]
manage = "manage:main"

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
Django = "^4.1.7"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"
django-cors-headers = "^3.14.0"
pillow = "^9.4.0"
django-filter = "^23.1"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipython = "^8.11.0"
ruff = "^0.0.259"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
    # Source folder
    src = ["src"] 

    exclude = [
        "*/migrations/*", 
        "manage.py",
    ]

    # Rules
    select = [  
        "F", 
        "I", 
        "N",
        "YTT",
        "S",
        "B",
        "COM",
        "C4",
        "DTZ",
        "DJ",
        "EM",
        "EXE",
        "ISC",
        "ICN",
        "G",
        "T20",
        "PYI",
        "Q",
        "RSE",
        "RET",
        "SIM",
        "PL",
    ]

    ignore = ["DJ001", "S107", "S106", "E501"] 
    target-version = "py311" # Python Version

